fixed bug

+++ PIE support Library/fit_data.m
@@ -5,13 +5,18 @@ function [pfit, ydata_fit, residuals, errfit] =fit_data(tdata,ydata,p0,str)
 % input: tdata, ydata,
 
 % check if a variable is to be fixed
-lb = str.lb;
-ub = str.ub;
+if length(str.lb)>=1
+    lb = str.lb;
+    ub = str.ub;
+end
+
 if str.remove_index > 0
     str.remove_xvalue=p0(str.remove_index);
     p0(str.remove_index)=[]; % shrink p0
-    lb(str.remove_index)=[];
-    ub(str.remove_index)=[];
+    if ~isEmpty(str.lb)
+        lb(str.remove_index)=[];
+        ub(str.remove_index)=[];
+    end
 end
 
 switch str.min_method
@@ -21,9 +26,8 @@ switch str.min_method
         options = optimoptions(@fminunc,'Algorithm','quasi-newton');
         [pfit,errfit] = fminunc(str.eval_function,p0,options); % tdata,ydata,str in global
     case 'chikv_optimize' % general  quasi-Newton method
-        options = optimset('Algorithm','sqp'); % we think we like sqp, but we aren't sure.
-        init_parray = p0; % this could be an input / randomized
-        [pfit, errfit] = fmincon(@(p)get_error(p,tdata,ydata,str), init_parray, [],[],[],[], lb, ub, [], options);
+        options = optimset('Algorithm','interior-point'); 
+        [pfit, errfit] = fmincon(@(p)get_error(p,tdata,ydata,str), p0, [],[],[],[], lb, ub, [], options);
         
  % tdata,ydata,str in global
     case 'MPP' % Moore-Penrose pseudo inverse for linear problems
